pyproject.toml
[tool.poetry]
name = "smtools"
version = "0.1.0"
description = ""
authors = ["hrsma2i <hrs.ma2i@gmail.com>"]
packages = [
    { include = "smtools" },
    { include = "sagemaker_tools" },
    { include = "sage_extensions" },
]

[tool.poetry.dependencies]
python = "^3.6"
joblib = "^0.14.1"
tqdm = "^4.41.1"
numpy = "1.16.4"
matplotlib = "^3.1.2"
pandas = "^0.25.3"
boto3 = "^1.11.0"
sagemaker = "^1.50.1"
slackweb = "^1.0.5"
chainer = "6.0.0"
pyyaml = "^5.3"
colorama = "^0.4.3"
pillow = "^6.2.1"
pytorch-ignite = "^0.3.0"

[tool.poetry.dev-dependencies]
mypy = "^0.761"
flake8 = "^3.7.9"
black = {version = "^19.10b0", allow-prereleases = true}
rope = "^0.16.0"
pytest = "^5.4.1"
jupyter = "^1.0.0"
ipython = "^7.13.0"
moto = "^1.3.14"
pytest-mock = "^3.0.0"

[tool.poetry.scripts]
smtrain = "sagemaker_tools.exec_train:main"
smdeploy = "sagemaker_tools.deploy_endpoint:main"
smbatch = "sagemaker_tools.batch_inference:main"
mgconf = "smtools.merge_configs:main"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
